requirements.txt
# Core dependencies
requests>=2.31.0
pyyaml>=6.0
boto3>=1.34.0
botocore>=1.34.0
pillow>=10.0.0
python-dateutil>=2.8.0

# Machine Learning
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
opencv-python>=4.8.0
numpy>=1.24.0

# Web UI (Streamlit)
streamlit>=1.28.0
pandas>=2.0.0

# Utilities
pathlib2>=2.3.0
secrets
urllib3>=2.0.0
warnings

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Optional: For advanced features
# transformers>=4.30.0
# diffusers>=0.20.0
# accelerate>=0.20.0
